Refactor stream module tests
Moved the before/after suite to the suite setup module, as it is a good
practice in Ginkgo framework.

Changed some sleepy waiting by an assertion to environment being "open".
It is better to assert that the state is what we expect, instead of
delaying the test start with a sleep. After the sleep, we are still
unsure about the state and it may not be correct. It is better to always
assert on expected state, if possible.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>