Refactor stream package (#150)
* Refactor socket module

Stop using atomic module, in favour of a pure mutex approach. Atomic has
a few caveats in different platforms e.g. ARMv64.

Minor tweak to the Makefile to allow other container technologies (e.g.
containerd) to start RabbitMQ container for testing.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>

* Refactor stream module tests

Moved the before/after suite to the suite setup module, as it is a good
practice in Ginkgo framework.

Changed some sleepy waiting by an assertion to environment being "open".
It is better to assert that the state is what we expect, instead of
delaying the test start with a sleep. After the sleep, we are still
unsure about the state and it may not be correct. It is better to always
assert on expected state, if possible.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>

* Update Makefile

Since we are using Ginkgo, it makes sense to use Ginkgo to run our
tests. We get the same options as with go test.

To use Ginkgo, some additional dependencies from Ginkgo are required.
Manually fetched those using go get to current version of Ginkgo.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>

* Update go routines in tests

Go routines that make Gomega assertions will panic if the assertion
fails. Ginkgo won't be able to capture the failed assertion without
deferring a GinkgoRecover() call in said go routines.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>