Tests and small refactors (#1)
* Add binary marshalling to Header type

This will be required to encode/decode binary frames from the wire.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Refactor Command interfaces

Make the function names Go idiomatic following Effective Go guidelines.
Command Read function should return and bubble up the error from
underlying binary package functions, so that we can react higher up in
the stack.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Add tests to PeerProperties command

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Remove Command interfaces tests

These tests were testing a fake implementation. It is best to test the
actual implementation, introducing a fake, where needed, to substitute a
real resource e.g. network, http server, etc.

By testing a fake, we are not adding code coverage, and we are aiming
for 100% coverage :-)

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Fix Header writing function

It was passing a reference to a function, instead of a pointer to a
fixed size integer.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Update TCP client to panic on error

Temporary measure to "throw errors to your face" until we determine a
proper way to handle errors coming from the wire, or from the decoder.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>