SASL authentication structs (#2)
* Refactor SASL struct and add tests for Write

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Add tests for SASL handshake struct

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Mark Peer Properties Write test as flaky

WriteMany function uses `range` to iterate over map types. This is
correct, however, `range` does not guarantee the order of iteration on
the items. Sometimes, the binary sequence won't be what we expect
because we set an expectation for specific order; however, we actually
don't care about the order, just that the elements are present. This is
hard to express when building a byte slice.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Add support structs for SASL auth

There's also a refactor to make wire formatting functions private. We do
not intend these functions to be used by users and they can be all used
from the internal package.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>