Open/Close connection (#3)
* Add Tune structs for frame encoding

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Add Open structs frames

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Add Close struct to encode frames

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Rename client to conform with other stream clients

Stream clients in other languages, like Java and .NET, use the term
'client' as a smart client, or a smart layer on top of a low level
client. We have agreed to keep consistent type names across the
different client implementations, so that low level client is 'raw' and
high level smart client is 'client'.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Fix read implementation

Some structs were flipping the most significant bit in the response
code. This is not correct because response codes do not change,
regardless of the sender. The most significant bit flip is meant for the
command code/key id decoding. This happens during the header decoding
and it is not required in the "body" part of the frame.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Add client configuration structs

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Add connection Connect

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Implement connection open/close

Connection open and close it tested using a fake connection via a pipe.
We mimic the read/write of frames that we expect RabbitMQ to
send/receive. In main.go, we use an actual RabbitMQ. This should be
removed later in the development cycle.

Decided to use go-logr as our logging solution because, as the author
very well explains, Go Core library does not have an interface for a
logger, therefore, it "forces" the developer to make a decision on the
logging technology to use. This is not correct in a library, because the
users of this library may have a different logging solution in-place,
and a client library should not dictate the logging technology of an
application; it should be the other way around.

Because of the above, go-logr is the best solution for logging, as it
provides an interface to a logger, and expects the actual implementation
to come from the user. We leverage the core library context package to
pass around a logger. If no logger is provided, we simply discard all
messages we intended to log. This will be documented later.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Drop govulncheck as a tool dependency

Because it requires to run on Go 1.18. We want to be able to develop
using Go 1.17.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Update frame writers to use wire formatting

We want to direct all buffer I/O operations via internal wire formatting
module.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Address feedback in #3

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>