Refactor raw package tests
Before this commit, a function to fake rabbitmq responses had to
hardcode the correlation ID in the request. This was very brittle and
would easily break if the order in which the fake functions was called.

Now there's a fake RabbitMQ server type, which has an auto-incrementing
sequence number. The server fakes now use this sequence to set the
correlation ID in responses from the fake RabbitMQ server. This removes
the strict order limitation for calling the fake functions.

Lifted 'n' shifted the fakes to the suite file. These functions will be
used in various part of the code. A place for shared types an functions
should be the suite.

Added an option to start `handleIncoming()` in the client, so that we
can test fakes without having to always call `Client.Connect()` and the
fakes required to make `Connect()` succeed.

Co-authored-by: Gabriele Santomaggio <gsantomaggio@vmware.com>
Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>