Refactors (#18)
* Refactor test functions and imports

Fake functions to mimic RabbitMQ responses were too rigid. Now they can
return custom response codes. This will be used to TDD the error
checking in the public API functions.

Had to move response codes to common package because this refactor was
creating a cyclic dependency as:

raw imports internal
-> internal imports common
-> -> common imports internal (cyclic)

Exposing the response codes to the user is acceptable, as we are going
to log the response codes, alongside with the error message.

Co-authored-by: Gabriele Santomaggio <gsantomaggio@vmware.com>
Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Refactor fake RabbitMQ functions

Remove duplication, making it easier to read the code.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Refactor fake Close

To have consistentcy with other fakes. Other fakes are using Header.Read
and the length of the frame to determine how much they have to read from
the buffer.

Co-authored-by: Gabriele Santomaggio <gsantomaggio@vmware.com>
Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>
Co-authored-by: Gabriele Santomaggio <gsantomaggio@vmware.com>