Handle close requests from server (#22)
* Refactor incoming frame handler

Incoming frame handler was decoding the command ID before dispatching to
the corresponding function. This is an issue when a command can be sent
from both the server and the client, because we can't distinguish from a
request and a response.

* Remove close response

Close response is a Simple Response. We can use simple response instead
and delete some code.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Update connection closure handler

The server can close a client connection from their end. Previously, we
were not handling the reception of a close request frame correctly.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

* Update Ginkgo to 2.5.1

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>