Rename DeclareConsumer to Subscribe
The protocol does not define a declare consumer frame. Declaring a
consumer would be better placed in the smart layer, where we abstract
some parts of the protocol, like subscribing and sending credits. In the
raw layer we should follow the stream protocol as close as possible.

Because of the above, DeclareConsumer function was renamed to Subscribe.

We were missing unit tests with fakes to test the delivery frame. Those
were added in this commit.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>