Fix incorrect buffer alignment on unknown frames
When we receive a frame not implemented e.g. Heartbeat, we read the
header and continue. This is incorrect, because we leave the body in the
network buffer. This will cause the next iteration of handleIncoming to
read the left-behind body, and interpret it as a header, which is not
correct, and results in misleading error messages. For example:

frame not implemented command ID=65321

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>