Delivery command version 2
- Implement version 2 of delivery
- Struct for 'exchange command versions' command
- Fix two bugs that surfaced after 8d38ca46383035e233600d2a788d4e43464213b2

One bug was in open.MarshalBinary. This function was writing the len of
the connection properties slice, before using writeMany to write the
connection properties. This was wrong because writeMany was changed to
write the len of a slice before writing the slice itself. This was
surfacing as an "unknown frame" error, because buffer reading was
misaligned.

The other bug was from implementing delivery V2, the amount of static
field changed, but a check in ChunkResponse.MarshalBinary was not
updated. This resulted in a nil,nil return, instead of returning the
actual byte slice.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>