Replace logr with slog
slog is coming to the standard library. The proposal is in accepted
state at the time of this writing. slog aims to solve the same problem
that logr solves: providing a front-end logger with a pluggable
back-end. This allows us, library developers, to simply "log"
information about what's happening inside the library, and keeps the
flexibility of app developers (consumers of this library), to decide and
configure a concrete logging implementation.

In addition, any error that is returned won't be logged by the library.
This is to avoid noisy logs, and potentially duplicating information.
Because most apps will emit some log when they receive an error, there's
no point in logging an error that we are returning (duplicate logging).

logr seems to be a dependency of Ginkgo V2, so we keep it in go-mod as
an indirect test dependency.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>