Change Raw Client configuration broker field
The Raw layer was keeping a reference to multiple URIs, but it was only
using the first one, always. We will handle the multiple URIs in the
smart layer. The raw layer will assume there is only one URI to connect
to, and the smart layer will have one client per rabbitmq node, and
decide which one to use.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>