Update handleIncoming in raw layer
The smart layer is not always receiving a closed notification from the
raw layer. I suspect we may not always enter the condition in
`tc.IsOpen()`, and the notification is not sent. I'm not able to
reproduce manually the problem after this patch.

In addition, we were losing the logger after the locator was
reconnecting, resulting in a discard logger after one reconnection. Now
it passes the logger used in the locator.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>