Fix verbosity logging in request
Removed the TODO related to checking context cancellation because
context cancellation is already checked in the very next line.

The TODO regarding a timeout in handleIncoming is discarded because it
is not safe to add a socket read timeout. We could add an idle deadline
to unblock from `header.Read()`, but we are unsure of what to expect in
the scenario where legitimate data arrives to the socket, and the
deadline is reached. We are concerned that we may get in a "half-read"
buffer situation. That would be catastrophic, since the next
`header.Read` will read from a "missaligned" buffer. We could "remember"
what we read and try to recompose the buffer in the next iteration, but
that feels like a lot of complexity for no tangible benefit (other than
not blocking this routine on `header.Read()`, which is not incorrect).

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>