Update handleIncoming FIXME's
We were returning an error in some places when we failed to decode a
frame body. Other clients don't do the same, and we prefer to keep going
and give a chance to the system to move on. If the buffer is
missaligned, the library will stop in the next iteration.

We changed the "default" or unknown frame to panic. We could end up here
if either we haven't implemented a handler for a specific frame, or the
buffer is read incorrectly and we can't determine whether next bytes are
part of the protocol, or a partially-read command. In either case, we
should panic. Earlier, we were assuming that default case meants an
unimplemented handler; that's not likely true at this stage.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>