Add support to determine server version
This is important to determine whether to send the exchange command
versions command, which is required to determine the version of certain
commands e.g. Deliver v1 or v2, Publisher v1 or v2. In the future,
changes in the protocol will be introduced with exchange command
versions, and we have to support sending this command for 3.11+

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>
Co-authored-by: Alex Blease <ablease@vmware.com>