Deprecate StreamOptions
It's an alias for backwards compatibility with v1 version of the client.
The type StreamOptions is very generic e.g. it could be used in delete
stream operations, however, it is specific to create stream operations.
The type name starts with the package name, which is not correct
according to Go best practices.

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>