Add dedicated type for message confirmation
We want to provide a better abstraction for message confirmations in the
smart layer. The publish confirm frame sends a potentially long list of
IDs that are confirmed. In the stream protocol, a confirm for ID N does
not confirm messages with IDs < N. Therefore, we have to keep track of
all pending publishing IDs, and associate them with their AMQP message.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>