Refactor message confirmation
At the raw layer, the client uses two different commands for publish
confirmation and publish error. At the smart layer, we abstract that
away and consider a publish error a "nack" or negative publish
confirmation. We have to adapt the channel in the producer, so that it
can receive publish confirmations or publish errors in the same channel.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>