MVP of smart producer (#255)
* Bump to Go 1.21
* Standard producer basic functionality
* First pass for smart producer.
* Make batch publishing delay configurable

* GA version of `slog` has a different signature for `slog.New`.
* This commit updates some code that was using the old signature.
* Smart Send: force send when buffer is full
* Publish confirmation tracker
* Add a semaphore to confirmation tracker

* The semaphore provides an simple mechanism to synchronise
the addition and release of pending confirmation
messages. Given how channels work in Go, the functions to
add pending confirmations will block if the semaphore
does not have more slots or permits available.

* Ground work for publisher confirms


* Refactor publisher message queue and confirm tracker

* Channels synchronise naturally, and also act as a synchronised
queue. Those two properties make channels an excellent choice
to implement a semaphore and a message queue buffer.

* Add dedicated type for message confirmation

* We want to provide a better abstraction for message confirmations in the
smart layer. The publish confirm frame sends a potentially long list of
IDs that are confirmed. In the stream protocol, a confirm for ID N does
not confirm messages with IDs < N. Therefore, we have to keep track of
all pending publishing IDs, and associate them with their AMQP message.

* To use a local binary version, and avoid changing the system binary for
Ginkgo CLI that may be present. The same for MockGen.

* Message confirmations in Send

* At the raw layer, the client uses two different commands for publish
confirmation and publish error. At the smart layer, we abstract that
away and consider a publish error a "nack" or negative publish
confirmation. We have to adapt the channel in the producer, so that it
can receive publish confirmations or publish errors in the same channel.

* Add unconfirmed messages in Send batch
* Confirm timeouts in standard producer
* Add Producer Manager component

* This component is responsible for keeping track of producers in the
environment. This component also sends the relevant frame to the broker,
so that the server registers a new publisher.
* Create Producer from Environment
* Use stable version of Go in CI

* We started developing in Go 1.21 because slog package became GA in Go
1.21. The toolchain directive in go.mod is new and only available in Go
1.21. We have to temporarily bump the Go version in CI to 'stable'. We
will provide compatibility to oldstable at a later date.

* Fix: delete publisher when producer is removed from manager
* Fix system test before suite

* Skipping system tests were not actually skipping system tests.

---------
Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>