Fix handling and storing offset by timestamp
Problem:

Sometimes the timestamp value can be written as an offset in rabbitmq, which freezes message consumption
Real case: You add a new consumer to an existing stream for which the offset is not yet set, and you do not want to reread all messages, so you specify -1 min. If there are no messages for a certain time and automatic flush is triggered, this timestamp value will be written as the offset.