Document manual credit flow with example (#434)
* Document manual credit flow

An example shows a somewhat complicated method to send credits every 10
chunks as manual strategy.

Refactored the ProducerOptions because there were some comments not
appearing in the GoDoc render. Comments in structs must be either inline
or above the field.

* Update examples readme

[skip ci]

* Fix linter error

Ignoring dogsled linter because we don't really care about the return
value of ReadLine, and we can't chante ReadLine code. We simply want to
block on that line until the user presses enter.