Rewrite from scracth manual credit test (#437)
The idea now is to test that eventually we don't receive more messages,
because we exhausted the credits, and that after issues enough credits,
we receive the last message in the batch. This test does not rely on
arbitrary or magic number of messages available per chunk, therefore it
should be more reliable.